requirements.txt
SQLAlchemy~=1.4.52
minio~=7.2.9
pydantic~=1.10.13
uvicorn~=0.31.0
fastapi~=0.109.1
selenium~=4.25.0
python-dotenv~=1.0.1
psycopg2